schwingbach: Added guard to prevent double imports in importlog.py Check in importpressure.py for start and enddate Added minvalue/maxvalue to ValueType Made Timeseries and TransformedTimeseries as children of Dataset. Added asarray function to dataset and DatasetGroup Added tools to create TransformedTimeseries in datasetpage.py and made load of statistics to ajax Added formatfloat to webpage/lib.py Simplified plotting by using asarray method Added tools to work with transformed datasets to datasettab.html

+++ schwingbach/db/dataset.py
@@ -12,7 +12,7 @@ from datetime import datetime,timedelta
 from dbobjects import newid, Person, Datasource
 from collections import deque
 from math import sqrt
-
+import numpy as np
 class Quality(Base):
     """Denotes the data quality of measurements, from raw to calibrated.
     id: numeric key to the dataquality
@@ -39,6 +39,8 @@ class ValueType(Base):
     name=sql.Column(sql.String)
     unit=sql.Column(sql.String)
     comment=sql.Column(sql.String)
+    minvalue=sql.Column(sql.Float)
+    maxvalue=sql.Column(sql.Float)
     def __str__(self):
         return "%s [%s]" % (self.name,self.unit)
     def __cmp__(self,other):
@@ -52,6 +54,9 @@ class ValueType(Base):
                     name=self.name,
                     unit=self.unit,
                     comment=self.comment)
+    def outofrange(self,value):
+        return ((self.minvalue is None or value>=self.minvalue)
+            and (self.maxvalue is None or value<=self.maxvalue))
 
        
     
@@ -102,7 +107,10 @@ class Dataset(Base):
     calibration_offset = sql.Column(sql.Float,nullable=False,default=0.0)
     calibration_slope = sql.Column(sql.Float,nullable=False,default=1.0)
     comment=sql.Column(sql.String)
-    
+    type=sql.Column(sql.String)
+    __mapper_args__ = dict(polymorphic_identity=None,
+                           polymorphic_on=type)
+
     def __str__(self):
         return (u'ds%(id)03i: %(valuetype)s at site #%(site)s with %(instrument)s (%(start)s-%(end)s)' % 
                 dict(id=self.id,
@@ -125,47 +133,12 @@ class Dataset(Base):
                     comment=self.comment,
                     label=self.__str__())
 
-    def maxrecordid(self):
-        """Finds the highest record id for this dataset"""
-        session = self.session()
-        q=session.query(sql.func.max(Record.id)).select_from(Record)
-        q=q.filter_by(dataset=self).scalar()
-        return q if not q is None else 1
-        
-    def addrecord(self,Id=None,value=None,time=None,comment=None):
-        """Adds a record to the dataset
-        Id: id for the recordset, if None, a new id will be created
-        value: value of the record
-        time: time of the record
-        comment: comments for this new record
-        """
-        value=float(value)
-        session = self.session()
-        if Id is None:
-            maxid = self.maxrecordid()
-            Id=maxid+1
-        if time is None:
-            time = datetime.now()
-        if not (self.start <= time <= self.end):
-            raise RuntimeError('RECORD does not fit DATASET: You tried to insert a record for date %s ' +
-                               'to dataset %s, which allows only records between %s and %s' 
-                               % (time,self,self.start,self.end))
-        result = Record(id=Id,time=time,value=value,dataset=self,comment=comment)
-        session.add(result)
-        return result
-    
-    def statistics(self):
-        """Calculates mean, stddev and number of records for this data set
-        """
-        session = self.session()
-        if not session:
-            return 0,0,0            
-        n = self.records.filter(Record.is_error == False).count()
-        if not n: return 0,0,0
-        mean = session.query(sql.func.avg(Record.value)).filter(Record.dataset==self).filter(Record.is_error == False).scalar()
-        std = session.query(sql.func.stddev(Record.value)).filter(Record.dataset==self).filter(Record.is_error == False).scalar()
-        return mean,std,n
-    
+    def is_timeseries(self):
+        return self.type=='timeseries'
+    def is_transformed(self):
+        return self.type=='transformed_timeseries'
+    def is_geodataset(self):
+        return self.type=='geodataset'
     def copy(self,id):
         """Creates a new dataset without records with the same meta data as this dataset.
         Give a new (unused) id
@@ -183,63 +156,15 @@ class Dataset(Base):
                         start=self.start,
                         end=self.end,
                         site=self.site)
-    
-    def split(self,time):
-        """Creates a new dataset using copy and assignes all records after
-        time to the new dataset. Useful """
-        session = self.session()
-        next = self.records.filter(Record.time>=time,Record.value != None).order_by(Record.time).first()
-        last = self.records.filter(Record.time<=time,Record.value != None).order_by(sql.desc(Record.time)).first()
-        if not next or not last:
-            raise RuntimeError("Split time %s is not between two records of %s" % (t,self))
-        
-        self.comment+='Dataset is splitted at %s to allow for different calibration' % time
-        dsnew = self.copy(id=newid(Dataset,session))
-        self.end = last.time
-        dsnew.start = next.time
-        records = self.records.filter(Record.time>=next.time)
-        records.update({'dataset':dsnew.id})
-        session.commit()
-        return self,dsnew
-        
-    
-    def findjumps(self,threshold):
-        """Returns an iterator to find all jumps greater than threshold
-        
-        To find "jumps", the records of the dataset are scanned for differences
-        between records (ordered by time). If the difference is greater than threshold,
-        the later record is returned as a jump
+    def asarray(self,start=None,end=None):
+        raise NotImplementedError('%s - data set can not return values with "asarray". Is the type correct?' % self.type)    
+    def size(self):
+        return 0
+    def statistics(self):
+        """Calculates mean, stddev and number of records for this data set
         """
-        records = self.records.order_by(Record.time).filter(Record.value != None,~Record.is_error)
-        last=None
-        for rec in records:
-            if not rec.value is None:
-                if last and abs(rec.value-last.value)>threshold:
-                    yield rec
-                last=rec
-    
-    def findvalue(self,time):
-        """Finds the linear interpolated value for the given time in the record"""
-        next = self.records.filter(Record.time>=time,Record.value != None,~Record.is_error).order_by(Record.time).first()
-        last = self.records.filter(Record.time<=time,Record.value != None,~Record.is_error).order_by(sql.desc(Record.time)).first()
-        if next and last:
-            dt_next = (next.time-time).total_seconds()
-            dt_last = (time-last.time).total_seconds()
-            dt = dt_next + dt_last
-            if dt<0.1:
-                return next.value,0.0
-            else:
-                return (1-dt_next/dt) * next.value + (1-dt_last/dt) * last.value,min(dt_last,dt_next)
-        elif next:
-            return next.value,(next.time-time).total_seconds()
-        elif last:
-            return last.value,(time-last.time).total_seconds()
-        else:
-            raise RuntimeError('%s has no records' % self)
-    def calibratevalue(self,value):
-        """Calibrates a value
-        """
-        return value * self.calibration_slope + self.calibration_offset
+        t,v = self.asarray()
+        return np.mean(v),np.std(v),len(v)
 
 def removedataset(*args):
     """Removes a dataset and its records entirely from the database
@@ -248,12 +173,24 @@ def removedataset(*args):
     datasets = [session.query(Dataset).get(int(a)) for a in args]
     for ds in datasets:
         name = str(ds)
-        reccount=ds.records.delete()
-        session.commit()
+        if ds.is_timeseries():
+            reccount=ds.records.delete()
+            session.commit()
+        else:
+            reccount = 0
         session.delete(ds)
         session.commit()
         print "Deleted %s and %i records" % (name,reccount)
 
+class MemRecord(object):
+    def __init__(self,id,dataset,time,value,sample=None,comment=None,is_error=False):
+        self.id=id
+        self.dataset=dataset
+        self.time=time
+        self.value=value
+        self.sample=sample
+        self.comment=comment
+        self.is_error=is_error
 class Record(Base):
     """
     The record holds sinigle measured, quantitative values.
@@ -270,7 +207,7 @@ class Record(Base):
     __tablename__= 'record'
     id=sql.Column(sql.Integer, primary_key=True)
     _dataset=sql.Column("dataset", sql.Integer, sql.ForeignKey('dataset.id'), primary_key=True)
-    dataset=orm.relationship("Dataset", backref= orm.backref('records',lazy='dynamic'),lazy='joined')
+    dataset=orm.relationship("Timeseries", backref= orm.backref('records',lazy='dynamic'),lazy='joined')
     time=sql.Column(sql.DateTime)
     value=sql.Column(sql.Float)
     sample=sql.Column(sql.String)
@@ -298,10 +235,185 @@ class Record(Base):
                     sample=self.sample,
                     comment=self.comment)
 
+# TODO: Put foreign_key to timeseries
+transforms_table = sql.Table('transforms',Base.metadata,
+                             sql.Column('target',sql.Integer,ForeignKey('transformed_timeseries.id'),primary_key=True),
+                             sql.Column('source',sql.Integer,ForeignKey('dataset.id'),primary_key=True))
+class Timeseries(Dataset):    
+    __mapper_args__ = dict(polymorphic_identity='timeseries')
+    def split(self,time):
+        """Creates a new dataset using copy and assignes all records after
+        time to the new dataset. Useful """
+        session = self.session()
+        next = self.records.filter(Record.time>=time,Record.value != None).order_by(Record.time).first()
+        last = self.records.filter(Record.time<=time,Record.value != None).order_by(sql.desc(Record.time)).first()
+        if not next or not last:
+            raise RuntimeError("Split time %s is not between two records of %s" % (t,self))
+        
+        self.comment+='Dataset is splitted at %s to allow for different calibration' % time
+        dsnew = self.copy(id=newid(Dataset,session))
+        self.end = last.time
+        dsnew.start = next.time
+        records = self.records.filter(Record.time>=next.time)
+        records.update({'dataset':dsnew.id})
+        session.commit()
+        return self,dsnew
+    def findjumps(self,threshold):
+        """Returns an iterator to find all jumps greater than threshold
+        
+        To find "jumps", the records of the dataset are scanned for differences
+        between records (ordered by time). If the difference is greater than threshold,
+        the later record is returned as a jump
+        """
+        records = self.records.order_by(Record.time).filter(Record.value != None,~Record.is_error)
+        last=None
+        for rec in records:
+            if not rec.value is None:
+                if last and abs(rec.value-last.value)>threshold:
+                    yield rec
+                last=rec
+    def findvalue(self,time):
+        """Finds the linear interpolated value for the given time in the record"""
+        next = self.records.filter(Record.time>=time,Record.value != None,~Record.is_error).order_by(Record.time).first()
+        last = self.records.filter(Record.time<=time,Record.value != None,~Record.is_error).order_by(sql.desc(Record.time)).first()
+        if next and last:
+            dt_next = (next.time-time).total_seconds()
+            dt_last = (time-last.time).total_seconds()
+            dt = dt_next + dt_last
+            if dt<0.1:
+                return next.value,0.0
+            else:
+                return (1-dt_next/dt) * next.value + (1-dt_last/dt) * last.value,min(dt_last,dt_next)
+        elif next:
+            return next.value,(next.time-time).total_seconds()
+        elif last:
+            return last.value,(time-last.time).total_seconds()
+        else:
+            raise RuntimeError('%s has no records' % self)
+    def calibratevalue(self,value):
+        """Calibrates a value
+        """
+        return value * self.calibration_slope + self.calibration_offset
+    def maxrecordid(self):
+        """Finds the highest record id for this dataset"""
+        session = self.session()
+        q=session.query(sql.func.max(Record.id)).select_from(Record)
+        q=q.filter_by(dataset=self).scalar()
+        return q if not q is None else 1
+    def addrecord(self,Id=None,value=None,time=None,comment=None):
+        """Adds a record to the dataset
+        Id: id for the recordset, if None, a new id will be created
+        value: value of the record
+        time: time of the record
+        comment: comments for this new record
+        """
+        value=float(value)
+        session = self.session()
+        if Id is None:
+            maxid = self.maxrecordid()
+            Id=maxid+1
+        if time is None:
+            time = datetime.now()
+        if (self.valuetype.outofrange(value)):
+            raise ValueError('RECORD does not fit VALUETYPE: %(v)g %(u)s is out of range for %(vt)s' 
+                             % dict(v=value,u=self.valuetype.unit,vt=self.valuetype.name)) 
+        if not (self.start <= time <= self.end):
+            raise ValueError('RECORD does not fit DATASET: You tried to insert a record for date %s ' +
+                               'to dataset %s, which allows only records between %s and %s' 
+                               % (time,self,self.start,self.end))
+        result = Record(id=Id,time=time,value=value,dataset=self,comment=comment)
+        session.add(result)
+        return result
+    def asarray(self,start=None,end=None):
+        session = self.session()    
+        records = session.query(Record).filter(Record._dataset == self.id)
+        records=records.order_by(Record.time).filter(~Record.is_error)
+        if start:
+            records=records.filter(Record.time>=start)
+        if end:
+            records=records.filter(Record.time<=end)
+        t0 = datetime(1,1,1)
+        date2num = lambda t: (t-t0).total_seconds()/86400 + 1.0
+        def r2c(records):
+            for r in records:
+                if r[0] is None:
+                    yield np.log(-1),date2num(r[1])
+                else:
+                    yield r[0],date2num(r[1])
+        t = np.zeros(shape=records.count(),dtype=float)
+        v = np.zeros(shape=records.count(),dtype=float)
+        for i,r in enumerate(r2c(records.values('value','time'))):
+            v[i],t[i] = r
+        v*=self.calibration_slope 
+        v+=self.calibration_offset  
+        return t,v
+    def size(self):
+        return self.records.count()
 
-                
+
+class TransformedTimeseries(Dataset):
+    __tablename__ = 'transformed_timeseries'
+    __mapper_args__ = dict(polymorphic_identity='transformed_timeseries')
+    id = sql.Column(sql.Integer,sql.ForeignKey('dataset.id'), primary_key=True)
+    expression = sql.Column(sql.String)
+    latex = sql.Column(sql.String)
+    sources = orm.relationship("Timeseries", secondary=transforms_table,order_by="Timeseries.start")
+    def sourceids(self):
+        return [s.id for s in self.sources]
+    def size(self):
+        return self.session().query(Record).filter(Record._dataset.in_(self.sourceids())).count()
+    def asarray(self,start=None,end=None):
+        src_t=[]
+        src_v=[]
+        for src in self.sources:
+            t,v = src.asarray(start,end)
+            src_t = np.concatenate((src_t,t))
+            src_v = np.concatenate((src_v,v))
+        src_v = self.transform(src_v)
+        return src_t, src_v
+    def updatetime(self):
+        self.start = min(ds.start for ds in self.sources)
+        self.end = max(ds.end for ds in self.sources)
+    def transform(self,x):
+        return eval(self.expression,{'x':x},np.__dict__)
+    def records(self, witherrors=False):
+        session = self.session()
+        srcrecords = session.query(Record).filter(Record._dataset.in_(self.sourceids())).order_by(Record.time)
+        if not witherrors:
+            srcrecords = srcrecords.filter(~Record.is_error)
+        i=0
+        for r in srcrecords:
+            i+=1
+            yield MemRecord(id=i,dataset=r.dataset,time=r.time,value=self.transform(r.calibrated),sample=r.sample,comment=r.comment)
+    def suitablesources(self):
+        session = self.session()
+        sourceids = self.sourceids()
+        res = session.query(Timeseries).filter(Timeseries.site==self.site,
+                                               ~Timeseries.id.in_(sourceids),
+                                               )
+        if len(self.sources):
+            vt = self.sources[0].valuetype
+            res = res.filter(Timeseries.valuetype == vt)
+        return res
+                  
+           
         
-        
-        
+class DatasetGroup(object):
+    def __init__(self,datasetids,start=None,end=None):
+        self.datasetids = list(datasetids)
+        self.start = start
+        self.end = end
+    def datasets(self,session):
+        return session.query(Dataset).filter(Dataset.id.in_(self.datasetids)).order_by(Dataset.start).all()
+    def asarray(self,session):
+        datasets=self.datasets(session)
+        src_t=[]
+        src_v=[]
+        for src in datasets:
+            t,v = src.asarray(self.start,self.end)
+            src_t = np.concatenate((src_t,t))
+            src_v = np.concatenate((src_v,v))
+        return src_t, src_v
+
